backend/requirements.txt
# Core Django
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9

# Authentication & Security
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0

# Development
gunicorn==21.2.0
whitenoise==6.6.0

# Testing (optional)
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# API Documentation
drf-spectacular==0.27.0

# Linting & Formatting (optional)
black==23.12.1
flake8==6.1.0
isort==5.13.2
